feat: add DevAvatar, DevTemplate, and DevTrial components; update DevSideBar links and routes
